setting and cover

+++ database/migrations/2023_03_27_013342_create_school_years_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->increments('id');
             $table->string('slug')->nullable();
             $table->string('name');
-            $table->integer('status')->default(2);
+            $table->tinyInteger('status')->default(2);
             $table->timestamps();
             $table->softDeletes();
         });
