+++ database/migrations/2023_03_27_044600_create_study_classes_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->string('name');
             $table->integer('id_major');
             $table->integer('id_level');
-            $table->integer('status')->default(1);
+            $table->tinyInteger('status')->default(1);
             $table->timestamps();
             $table->softDeletes();
         });
